requirements-streaming.txt
# Kafka streaming requirements
kafka-python==2.0.2
psycopg2-binary==2.9.7
python-dateutil==2.8.2

# Additional utilities
requests==2.31.0
python-json-logger==2.0.7